Try properly rename/putting of the files.  Reset rename information to get the list and return the object if it doesn't find the data.
git-svn-id: https://svn.code.sf.net/p/freeghost/code/trunk@5213 71f96598-fa45-0410-b640-bcd6f8691b32

+++ packages/web/service/Post_Stage2.php
@@ -20,7 +20,7 @@ try {
         ->set('password',$StorageNode->get('pass'))
         ->connect()
         ->delete($dest);
-    if (!$FOGFTP->rename($src,$dest) && !$FOGFTP->put($dest,$src)) throw new Exception(_('Move/rename failed'));
+    if (!($FOGFTP->rename($src,$dest) || $FOGFTP->put($dest,$src))) throw new Exception(_('Move/rename failed'));
     in_array($_REQUEST['osid'],array(1,2)) ? $FOGFTP->delete(sprintf('%s/dev/%s',$StorageNode->get('ftppath'),$macftp)) : null;
     $FOGFTP->chmod(0755,$dest);
     $FOGFTP->close();
